\name{bioChemists}
\alias{bioChemists}
\docType{data}
\title{article production by graduate students in biochemistry Ph.D. programs}
\description{
  A sample of 915 biochemistry graduate students.
}
\usage{data(bioChemists)}
\format{
  \describe{
    \item{\code{art}}{count of articles produced during last 3 years of Ph.D.}
    \item{\code{fem}}{factor indicating gender of student, with levels Men and Women}
    \item{\code{mar}}{factor indicating marital status of student, with levels
      Single and Married}
    \item{\code{kid5}}{number of children aged 5 or younger}
    \item{\code{phd}}{prestige of Ph.D. department}
    \item{\code{ment}}{count of articles produced by Ph.D. mentor during last 3 years}
    }
  }
%%\source{found in Stata format at \url{http://www.indiana.edu/~jslsoc/stata/socdata/couart2.dta}}
\references{Long, J. Scott. 1990.  The origins of sex differences in
  science. \emph{Social Forces}. 68(3):1297-1316.

  Long, J. Scott. 1997.  \emph{Regression Models for Categorical and
    Limited Dependent Variables}. Thousand Oaks, California: Sage.
}
\keyword{datasets}
