\name{prussian}
\alias{prussian}
\docType{data}
\title{Prussian Army Horse Kick Data}
\description{
  Deaths by year, by corp, from horse kicks.
}
\usage{data(prussian)}
\format{
  A data frame with 20 observations on the following 14 variables.  Each
  variable contains the annual counts for a corp.
  \describe{
    \item{\code{G}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\code{II}}{a numeric vector}
    \item{\code{III}}{a numeric vector}
    \item{\code{IV}}{a numeric vector}
    \item{\code{V}}{a numeric vector}
    \item{\code{VI}}{a numeric vector}
    \item{\code{VII}}{a numeric vector}
    \item{\code{VIII}}{a numeric vector}
    \item{\code{IX}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{XI}}{a numeric vector}
    \item{\code{XIV}}{a numeric vector}
    \item{\code{XV}}{a numeric vector}
  }
}
\source{
  von Bortkiewicz, L. 1898. \emph{Das Gesetz der Kleinen Zahlen.} Leipzig: Teubner. 
}
\keyword{datasets}
